Add is_optional handling to PandasColumn so that we have optional columns schemas.
Summary: SQL schemas support the insertions of rows that have missing columns. We should make this functionality available. However, doing so leads to a lot of awkwardness with the `ColumnExistsConstraint` which has always been an awkward constraint. Aren't all constraints reliant on the fact that a column exists? It also makes the whole interface fragile because it HAD to be applied first before other constraints. Given all of this, this revision gets rid of the `ColumnExistsConstraint` in favor of using the existance of the column and an `is_optional` field as control logic in the `PandasColumn` validation function. This is a breaking change if consumers were using the low level API and using the `ColumnExistsConstraint`.

Test Plan: unit

Differential Revision: https://dagster.phacility.com/D2234